test(desktop): unify buy/sell navigation tests with describe.each
