+++ .changeset/thin-falcons-shop.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+unify buy / sell tests on market
